Removed studentcourse interface, added css modules, restructured component to match
