+++ src/features/dashboard/StudentDashboardSection.tsx
@@ -1,10 +1,8 @@
 import React from 'react';
-import { StudentCourseCard } from '../student-course-card/StudentCourseCard';
+import { StudentCourseCard } from '../shared/components/student-course-card/StudentCourseCard';
 
 export const StudentDashboardSection: React.FC = () => (
   <div>
-    {/* Student-specific content goes here */}
-    <p>Welcome, student! This is your dashboard.</p>
     <StudentCourseCard />
   </div>
 );
