+++ src/features/shared/components/student-course-card/StudentCourseCard.module.css
